+++ src/main/java/ru/hh/resumebuilderbot/BotBodyImpl.java
@@ -14,9 +14,9 @@ public class BotBodyImpl implements BotBody {
     public void answer(User user, Answer answer) {
         synchronized (UserDataStorage.getMutex(user)) {
             MessageHandler messageHandler = Selector.select(answer);
-            QuestionsGenerator questionGenerator = messageHandler.handle(user, answer);
-            Queue<Question> questions = questionGenerator.generateQuestions();
-            questions.forEach((x) -> messengerAdapter.ask(user, x));
+            QuestionsGenerator questionsGenerator = messageHandler.handle(user, answer);
+            Queue<Question> questions = questionsGenerator.generateQuestions();
+            questions.forEach((question) -> messengerAdapter.ask(user, question));
         }
     }
