+++ src/main/java/ru/hh/resumebuilderbot/MessengerAdapter.java
@@ -5,7 +5,5 @@ import ru.hh.resumebuilderbot.question.Question;
 public interface MessengerAdapter {
     void ask(User user, Question question);
 
-    void setBotBody(BotBody botBody);
-
     void start();
 }
