+++ src/main/java/ru/hh/resumebuilderbot/telegram/adapter/TelegramAdapter.java
@@ -2,59 +2,47 @@ package ru.hh.resumebuilderbot.telegram.adapter;
 
 import org.telegram.telegrambots.TelegramBotsApi;
 import org.telegram.telegrambots.api.methods.send.SendMessage;
-import org.telegram.telegrambots.api.objects.Update;
 import org.telegram.telegrambots.api.objects.replykeyboard.InlineKeyboardMarkup;
+import org.telegram.telegrambots.api.objects.replykeyboard.ReplyKeyboardMarkup;
+import org.telegram.telegrambots.api.objects.replykeyboard.ReplyKeyboardRemove;
 import org.telegram.telegrambots.api.objects.replykeyboard.buttons.InlineKeyboardButton;
+import org.telegram.telegrambots.api.objects.replykeyboard.buttons.KeyboardButton;
+import org.telegram.telegrambots.api.objects.replykeyboard.buttons.KeyboardRow;
 import org.telegram.telegrambots.bots.TelegramLongPollingBot;
 import org.telegram.telegrambots.exceptions.TelegramApiException;
-import ru.hh.resumebuilderbot.Answer;
 import ru.hh.resumebuilderbot.BotBody;
 import ru.hh.resumebuilderbot.MessengerAdapter;
 import ru.hh.resumebuilderbot.User;
 import ru.hh.resumebuilderbot.question.Question;
-import ru.hh.resumebuilderbot.telegram.adapter.answer.TelegramAnswer;
-import ru.hh.resumebuilderbot.telegram.adapter.answer.TelegramAnswerFactory;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class TelegramAdapter implements MessengerAdapter {
-    private final String token;
-    private final String botUsername;
-    private final BotImpl bot;
+    private final TelegramLongPollingBot bot;
 
-    private BotBody botBody;
-
-    public TelegramAdapter(String token, String botUsername) {
-        this.token = token;
-        this.botUsername = botUsername;
-        this.bot = new BotImpl();
-    }
-
-    @Override
-    public void setBotBody(BotBody botBody) {
-        this.botBody = botBody;
+    public TelegramAdapter(TelegramLongPollingBot bot) {
+        this.bot = bot;
     }
 
     @Override
     public void ask(User user, Question question) {
-        SendMessage msg = new SendMessage()
+        SendMessage message = new SendMessage()
                 .setChatId(user.getIndex())
                 .setText(question.getText());
 
         List<String> variantsOfAnswer = question.getVariantsOfAnswer();
-
         if (!variantsOfAnswer.isEmpty()) {
-            msg.setReplyMarkup(generateMarkup(variantsOfAnswer));
-        }
+            message.setReplyMarkup(generateInlineKeyboard(variantsOfAnswer));
+        } 
         try {
-            bot.sendMsg(msg);
+            bot.sendMessage(message);
         } catch (TelegramApiException e) {
             e.printStackTrace();
         }
     }
 
-    private InlineKeyboardMarkup generateMarkup(List<String> variantsOfAnswer) {
+    private InlineKeyboardMarkup generateInlineKeyboard(List<String> variantsOfAnswer) {
         InlineKeyboardMarkup result = new InlineKeyboardMarkup();
         List<List<InlineKeyboardButton>> rowsInline = new ArrayList<>();
 
@@ -77,36 +65,6 @@ public class TelegramAdapter implements MessengerAdapter {
             botsApi.registerBot(bot);
         } catch (TelegramApiException e) {
             e.printStackTrace();
-        }
-    }
-
-    private class BotImpl extends TelegramLongPollingBot {
-        void sendMsg(SendMessage msg) throws TelegramApiException {
-            sendMessage(msg);
-        }
-
-        @Override
-        public void onUpdateReceived(Update update) {
-            TelegramAnswer telegramAnswer = TelegramAnswerFactory.create(update);
-            if (telegramAnswer != null) {
-
-
-                long chatId = telegramAnswer.getChatId();
-                User user = new User(chatId);
-                String answerText = telegramAnswer.getAnswerText();
-                Answer answer = new Answer(answerText);
-                botBody.answer(user, answer);
-            }
-        }
-
-        @Override
-        public String getBotUsername() {
-            return botUsername;
-        }
-
-        @Override
-        public String getBotToken() {
-            return token;
         }
     }
 }
