+++ src/test/java/ru/hh/resumebuilderbot/testmessengeradapter/TestMessengerAdapter.java
@@ -31,10 +31,6 @@ public class TestMessengerAdapter implements MessengerAdapter {
         bot.answer(user, getAnswer(question));
     }
 
-    @Override
-    public void setBotBody(BotBody handler) {
-
-    }
 
     @Override
     public void start() {
